h5: moved all DD code to gridstack-dd.ts
* all drag&drop grid code has been moved to gridstack-dd.ts (base class for html5 and jquery version)
* that reduces static grid requirements down from 43k to just 35k
* with html5 complete is still 64k (+29k) and jq 187k (+152k)
