+++ src/index-h5.ts
@@ -5,7 +5,7 @@
 export * from './types';
 export * from './utils';
 export * from './gridstack-engine';
-export * from './gridstack-dd';
+export * from './gridstack-ddi';
 export * from './gridstack';
 
 export * from './h5/gridstack-dd-native';
